requirements.txt
fastapi==0.120.0
uvicorn==0.38.0
openai==2.6.1
python-dotenv==1.1.1
requests==2.32.5
langchain==1.0.2
langchain-core==1.0.0
langgraph==1.0.1
langsmith==0.4.38
sentence-transformers==5.1.2
pinecone==7.3.0
transformers==4.57.1
torch==2.9.0
pydantic==2.12.3
numpy==2.3.4
scikit-learn==1.7.2
PyMuPDF==1.26.5  